core: expand parallel state megatron: use flatflow parallel state
