+++ flatflow/megatron/core/__init__.py
@@ -1,0 +1,7 @@
+from flatflow.megatron.core.parallel_state import (
+    get_model_parallel_src_rank
+)
+
+__all__ = [
+    "get_model_parallel_src_rank",
+]
